refactor(atomic): share the Vite plugins that serve Atomic from source
Storybook and @coveo/atomic-playground both serve Atomic from `src` rather than
`dist`, and both had their own copy of the five plugins that reproduce the build's
transforms: the `@/` alias, the `custom-element-tags` stand-in, `?inline` CSS
imports, Tailwind directives inside Lit `css` templates, and inlined SVG imports.
These encode Atomic's build semantics, so two copies meant a component could
render correctly in one and not the other.

Keeps the playground's variants, which scope the transforms to Atomic source
instead of every `.ts` in the graph, normalize paths for Windows, preserve source
maps on files they leave untouched, and drop an `any`. No `.ts` outside
`packages/atomic/src` imports `.css` or `.svg` or writes Tailwind directives in
tagged templates, so the narrower scope is equivalent for Storybook.

One incidental change: Storybook no longer emits a `.shadow` utility that Tailwind
had been generating from the word "shadow" in a prose doc comment. No component
uses that class.
