+++ packages/atomic-playground/vite/atomic-source-plugins.ts
@@ -1,21 +1,14 @@
-import {readFileSync} from 'node:fs';
-import {dirname, resolve} from 'node:path';
+import {resolve} from 'node:path';
 import tailwindcss from '@tailwindcss/vite';
-import type {AcceptedPlugin} from 'postcss';
 import {normalizePath, type Plugin, type PluginOption} from 'vite';
-import {atomicPackageRoot, resolveToSource} from './source-mappings.js';
+import {
+  atomicPackageRoot,
+  atomicSourceTransformPlugins,
+} from '../../atomic/scripts/vite-source-plugins.js';
+import {resolveToSource} from './source-mappings.js';
 
 const atomicSrc = normalizePath(resolve(atomicPackageRoot, 'src'));
 const atomicTailwindEntry = normalizePath(resolve(atomicSrc, 'utils/tailwind.global.tw.css'));
-
-/**
- * Module ids arrive from Vite with POSIX separators even on Windows, so both sides of the
- * comparison are normalized before matching.
- */
-const isAtomicSource = (id: string) => {
-  const cleanId = normalizePath(id.split('?')[0]);
-  return cleanId.endsWith('.ts') && cleanId.startsWith(atomicSrc);
-};
 
 /**
  * Redirects workspace package specifiers to their source entry points.
@@ -23,6 +16,9 @@ const isAtomicSource = (id: string) => {
  * Serving Atomic and Headless from source is what lets Vite own the whole dependency graph:
  * an edit anywhere in either package invalidates exactly the affected modules, with no
  * intermediate `dist` rebuild to wait on and no filesystem watcher to keep in sync.
+ *
+ * Storybook does the same thing through its own `externalizeDependencies`, which additionally
+ * has to choose between source and CDN depending on the build mode.
  */
 const resolveWorkspaceSource = (): Plugin => ({
   name: 'resolve-workspace-source',
@@ -34,139 +30,13 @@ const resolveWorkspaceSource = (): Plugin => ({
 });
 
 /**
- * Resolves the `@/` alias Atomic source uses for imports outside the current directory.
- */
-const resolveAtomicPathAliases = (): Plugin => ({
-  name: 'resolve-atomic-path-aliases',
-  enforce: 'pre',
-  async resolveId(source, importer, options) {
-    if (source.startsWith('@/')) {
-      return this.resolve(resolve(atomicPackageRoot, source.slice(2)), importer, options);
-    }
-    return null;
-  },
-});
-
-/**
- * Replaces `src/utils/custom-element-tags.ts` with a module derived from the component map.
- *
- * That file is generated by `build:lit`, which the playground deliberately does not run.
- */
-const virtualCustomElementTags = (): Plugin => {
-  const virtualId = 'virtual:custom-element-tags';
-  const generatedFile = resolve(atomicSrc, 'utils/custom-element-tags.js');
-
-  return {
-    name: 'virtual-custom-element-tags',
-    enforce: 'pre',
-    resolveId(id, importer) {
-      if (importer && resolve(dirname(importer), id) === generatedFile) {
-        return virtualId;
-      }
-      return null;
-    },
-    load(id) {
-      if (id === virtualId) {
-        return `
-          import elementMap from '@/src/components/lazy-index.js';
-          export const ATOMIC_CUSTOM_ELEMENT_TAGS = new Set(Object.keys(elementMap));
-        `;
-      }
-      return null;
-    },
-  };
-};
-
-/**
- * Makes `.css` imports in Atomic source yield strings rather than injected stylesheets, so
- * they can be handed to Lit's `css` tag and adopted by a shadow root.
- */
-const forceInlineCssImports = (): Plugin => ({
-  name: 'force-inline-css-imports',
-  enforce: 'pre',
-  transform(code, id) {
-    if (!isAtomicSource(id)) {
-      return null;
-    }
-    const transformed = code.replace(
-      /import\s+([^'"]+)\s+from\s+['"]([^'"]+\.css)['"]/g,
-      (_, importName, cssPath) => `import ${importName} from '${cssPath}?inline'`
-    );
-    // Returning a result for a file this left untouched would discard its source map.
-    if (transformed === code) {
-      return null;
-    }
-    return {code: transformed, map: null};
-  },
-});
-
-/**
- * Resolves Tailwind directives inside Lit `css` tagged template literals.
- *
- * Components write `@import`, `@reference` and `@apply` in tagged `css` templates for their
- * shadow DOM styles. Production builds post-process those through `litCssPlugin`; here the
- * same work has to happen before the CSS reaches the browser, since constructable
- * stylesheets do not understand the directives.
- */
-const processInlineCssImports = (): Plugin => {
-  let postcssPlugins: AcceptedPlugin[];
-  let initialized = false;
-
-  return {
-    name: 'process-inline-css-imports',
-    enforce: 'pre',
-    async transform(code, id) {
-      if (!isAtomicSource(id)) {
-        return null;
-      }
-
-      const matches = [...code.matchAll(/css\s*`([\s\S]*?)`/g)];
-      const tailwindMatches = matches.filter(
-        (match) =>
-          match[1].includes('@import') ||
-          match[1].includes('@reference') ||
-          match[1].includes('@apply')
-      );
-      if (tailwindMatches.length === 0) {
-        return null;
-      }
-
-      if (!initialized) {
-        const {default: tailwindPostcss} = await import('@tailwindcss/postcss');
-        postcssPlugins = [tailwindPostcss()];
-        initialized = true;
-      }
-
-      const {default: postcss} = await import('postcss');
-      let result = code;
-
-      for (const match of tailwindMatches.reverse()) {
-        const processed = await postcss(postcssPlugins).process(match[1], {from: id});
-        // Escape backticks and template expression markers so the processed CSS can safely
-        // live inside a tagged template literal.
-        const escaped = processed.css
-          .replace(/\\/g, '\\\\')
-          .replace(/`/g, '\\`')
-          .replace(/\$\{/g, '\\${');
-        result =
-          result.slice(0, match.index) +
-          `css\`${escaped}\`` +
-          result.slice(match.index! + match[0].length);
-      }
-
-      return {code: result, map: null};
-    },
-  };
-};
-
-/**
  * Points Tailwind's content scanning at Atomic source.
  *
  * `@tailwindcss/vite` infers what to scan from the Vite root, which here holds only the
- * playground's own two files. Components spell out utility classes in their Lit templates and
+ * playground's own pages. Components spell out utility classes in their Lit templates and
  * rely on the global sheet `withTailwindStyles` adopts, so without this the `utilities` layer
- * comes out nearly empty and components render unstyled. The production build avoids the problem
- * by running PostCSS from the Atomic package directory.
+ * comes out nearly empty and components render unstyled. Storybook and the production build
+ * avoid the problem by running from the Atomic package directory, so this stays playground-only.
  */
 const scanAtomicSourceForUtilities = (): Plugin => ({
   name: 'scan-atomic-source-for-utilities',
@@ -180,33 +50,6 @@ const scanAtomicSourceForUtilities = (): Plugin => ({
 });
 
 /**
- * Inlines `.svg` imports as strings, matching how the production build handles them.
- */
-const svgTransform = (): Plugin => ({
-  name: 'svg-transform',
-  transform(code, id) {
-    if (!isAtomicSource(id)) {
-      return null;
-    }
-    const transformed = code.replace(
-      /import\s+([a-zA-Z]+)\s+from\s+['"]([^'"]+\.svg)['"]/g,
-      (_, importName, importPath) => {
-        const svgContent = readFileSync(resolve(dirname(id), importPath), 'utf8')
-          .replace(/\r?\n/g, '')
-          .replace(/\\/g, '\\\\')
-          .replace(/'/g, "\\'");
-        return `const ${importName} = '${svgContent}';`;
-      }
-    );
-    // Returning a result for a file this left untouched would discard its source map.
-    if (transformed === code) {
-      return null;
-    }
-    return {code: transformed, map: null};
-  },
-});
-
-/**
  * Vite plugins required to run Atomic and Headless from source.
  *
  * Component edits resolve to a full page reload rather than a patch, because
@@ -216,11 +59,7 @@ const svgTransform = (): Plugin => ({
  */
 export const atomicSourcePlugins = (): PluginOption[] => [
   resolveWorkspaceSource(),
-  resolveAtomicPathAliases(),
-  virtualCustomElementTags(),
-  forceInlineCssImports(),
-  processInlineCssImports(),
+  ...atomicSourceTransformPlugins(),
   scanAtomicSourceForUtilities(),
-  svgTransform(),
   tailwindcss(),
 ];
