+++ packages/atomic/scripts/vite-source-plugins.ts
@@ -1,0 +1,197 @@
+import {readFileSync} from 'node:fs';
+import {dirname, resolve} from 'node:path';
+import type {AcceptedPlugin} from 'postcss';
+import {normalizePath, type Plugin} from 'vite';
+
+/**
+ * Vite plugins that reproduce, in a dev server, the transforms Atomic's production build
+ * applies to its own source.
+ *
+ * Anything that serves Atomic from `src` rather than `dist` needs all of these: Storybook in
+ * development and Chromatic modes, and `@coveo/atomic-playground`. They live here so the two
+ * consumers cannot drift apart — these transforms encode Atomic's build semantics, so a
+ * component that renders correctly in one has to render correctly in the other.
+ */
+
+export const atomicPackageRoot = normalizePath(resolve(import.meta.dirname, '..'));
+
+const atomicSrc = normalizePath(resolve(atomicPackageRoot, 'src'));
+
+/**
+ * Module ids arrive from Vite with POSIX separators even on Windows, so both sides of the
+ * comparison are normalized before matching.
+ *
+ * Only Atomic's own source needs these transforms. Headless, Bueno, Relay and
+ * `@coveo/platform-mock-api` are also served from source, and none of them import `.css` or
+ * `.svg` or write Tailwind directives in tagged templates.
+ */
+const isAtomicSource = (id: string) => {
+  const cleanId = normalizePath(id.split('?')[0]);
+  return cleanId.endsWith('.ts') && cleanId.startsWith(atomicSrc);
+};
+
+/**
+ * Resolves the `@/` alias Atomic source uses for imports outside the current directory.
+ */
+const resolveAtomicPathAliases = (): Plugin => ({
+  name: 'resolve-atomic-path-aliases',
+  enforce: 'pre',
+  async resolveId(source, importer, options) {
+    if (source.startsWith('@/')) {
+      return this.resolve(resolve(atomicPackageRoot, source.slice(2)), importer, options);
+    }
+    return null;
+  },
+});
+
+/**
+ * Replaces `src/utils/custom-element-tags.ts` with a module derived from the component map.
+ *
+ * That file is generated by `build:lit`, which neither consumer runs.
+ */
+const virtualCustomElementTags = (): Plugin => {
+  const virtualId = 'virtual:custom-element-tags';
+  const generatedFile = resolve(atomicSrc, 'utils/custom-element-tags.js');
+
+  return {
+    name: 'virtual-custom-element-tags',
+    enforce: 'pre',
+    resolveId(id, importer) {
+      if (importer && resolve(dirname(importer), id) === generatedFile) {
+        return virtualId;
+      }
+      return null;
+    },
+    load(id) {
+      if (id === virtualId) {
+        return `
+          import elementMap from '@/src/components/lazy-index.js';
+          export const ATOMIC_CUSTOM_ELEMENT_TAGS = new Set(Object.keys(elementMap));
+        `;
+      }
+      return null;
+    },
+  };
+};
+
+/**
+ * Makes `.css` imports in Atomic source yield strings rather than injected stylesheets, so
+ * they can be handed to Lit's `css` tag and adopted by a shadow root.
+ */
+const forceInlineCssImports = (): Plugin => ({
+  name: 'force-inline-css-imports',
+  enforce: 'pre',
+  transform(code, id) {
+    if (!isAtomicSource(id)) {
+      return null;
+    }
+    const transformed = code.replace(
+      /import\s+([^'"]+)\s+from\s+['"]([^'"]+\.css)['"]/g,
+      (_, importName, cssPath) => `import ${importName} from '${cssPath}?inline'`
+    );
+    // Returning a result for a file this left untouched would discard its source map.
+    if (transformed === code) {
+      return null;
+    }
+    return {code: transformed, map: null};
+  },
+});
+
+/**
+ * Resolves Tailwind directives inside Lit `css` tagged template literals.
+ *
+ * Components write `@import`, `@reference` and `@apply` in tagged `css` templates for their
+ * shadow DOM styles. Production builds post-process those through `litCssPlugin`; here the
+ * same work has to happen before the CSS reaches the browser, since constructable
+ * stylesheets do not understand the directives.
+ */
+const processInlineCssImports = (): Plugin => {
+  let postcssPlugins: AcceptedPlugin[];
+  let initialized = false;
+
+  return {
+    name: 'process-inline-css-imports',
+    enforce: 'pre',
+    async transform(code, id) {
+      if (!isAtomicSource(id)) {
+        return null;
+      }
+
+      const matches = [...code.matchAll(/css\s*`([\s\S]*?)`/g)];
+      const tailwindMatches = matches.filter(
+        (match) =>
+          match[1].includes('@import') ||
+          match[1].includes('@reference') ||
+          match[1].includes('@apply')
+      );
+      if (tailwindMatches.length === 0) {
+        return null;
+      }
+
+      if (!initialized) {
+        const {default: tailwindPostcss} = await import('@tailwindcss/postcss');
+        postcssPlugins = [tailwindPostcss()];
+        initialized = true;
+      }
+
+      const {default: postcss} = await import('postcss');
+      let result = code;
+
+      for (const match of tailwindMatches.reverse()) {
+        const processed = await postcss(postcssPlugins).process(match[1], {from: id});
+        // Escape backticks and template expression markers so the processed CSS can safely
+        // live inside a tagged template literal.
+        const escaped = processed.css
+          .replace(/\\/g, '\\\\')
+          .replace(/`/g, '\\`')
+          .replace(/\$\{/g, '\\${');
+        result =
+          result.slice(0, match.index) +
+          `css\`${escaped}\`` +
+          result.slice(match.index! + match[0].length);
+      }
+
+      // Expanding `@apply` shifts every position after the template, so the incoming source
+      // map cannot be carried through.
+      return {code: result, map: null};
+    },
+  };
+};
+
+/**
+ * Inlines `.svg` imports as strings, matching how the production build handles them.
+ */
+const svgTransform = (): Plugin => ({
+  name: 'svg-transform',
+  transform(code, id) {
+    if (!isAtomicSource(id)) {
+      return null;
+    }
+    const transformed = code.replace(
+      /import\s+([a-zA-Z]+)\s+from\s+['"]([^'"]+\.svg)['"]/g,
+      (_, importName, importPath) => {
+        const svgContent = readFileSync(resolve(dirname(id), importPath), 'utf8')
+          .replace(/\r?\n/g, '')
+          .replace(/\\/g, '\\\\')
+          .replace(/'/g, "\\'");
+        return `const ${importName} = '${svgContent}';`;
+      }
+    );
+    // Returning a result for a file this left untouched would discard its source map.
+    if (transformed === code) {
+      return null;
+    }
+    return {code: transformed, map: null};
+  },
+});
+
+/**
+ * The transforms every consumer that serves Atomic from source needs, in dependency order.
+ */
+export const atomicSourceTransformPlugins = (): Plugin[] => [
+  resolveAtomicPathAliases(),
+  virtualCustomElementTags(),
+  forceInlineCssImports(),
+  processInlineCssImports(),
+  svgTransform(),
+];
